Notes try to open in app, fallback to web
Also added a BOOL that says if Evernote has fixed the viewNoteInEvernote problem yet (sync happens too late, so the new note can't be found). Once that works correctly on Evernote's side, I'll just have to change the value to true and the specific note will be opened, not just the app. The banner text will also reflect that difference.